cu29_derive

Attribute Macro copper_runtime

#[copper_runtime]

Adds #[copper_runtime(config = "path", subsystem = "id", sim_mode = false/true, ignore_resources = false/true)] to your application struct to generate the runtime. if sim_mode is omitted, it is set to false. if ignore_resources is omitted, it is set to false. if subsystem is provided, config must point to a strict multi-Copper config and the selected subsystem local config will be embedded into the generated runtime. This will add a “runtime” field to your struct and implement the “new” and “run” methods.